requirements.txt
numpy==1.22.3
pandas==1.4.2
Pillow==9.2.0
streamlit==1.9.0
sklearn
protobuf==3.20
